requirements.txt
-i https://pypi.org/simple
altair==4.2.2 ; python_version >= '3.7'
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2 ; python_version >= '3.7'
bs4==0.0.1
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1 ; python_version >= '3.5'
entrypoints==0.4 ; python_version >= '3.6'
exceptiongroup==1.1.1
fonttools==4.39.4
gitdb==4.0.10 ; python_version >= '3.7'
gitpython==3.1.31 ; python_version >= '3.7'
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
joblib==1.2.0
jsonschema==4.17.3 ; python_version >= '3.7'
kiwisolver==1.4.4
markdown-it-py==2.2.0 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
matplotlib==3.7.1
mdurl==0.1.2 ; python_version >= '3.7'
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pillow==9.5.0
protobuf==3.20.3 ; python_version >= '3.7'
pyarrow==12.0.0 ; python_version >= '3.7'
pycparser==2.21
pydeck==0.8.1b0 ; python_version >= '3.7'
pygments==2.15.1 ; python_version >= '3.7'
pympler==1.0.1 ; python_version >= '3.6'
pyparsing==3.0.9
pyrsistent==0.19.3 ; python_version >= '3.7'
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
rich==13.3.5 ; python_full_version >= '3.7.0'
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.9.1
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0 ; python_version >= '3.6'
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
streamlit==1.22.0
tabulate==0.9.0
tenacity==8.2.2 ; python_version >= '3.6'
threadpoolctl==3.1.0
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.12.0 ; python_version >= '3.5'
tornado==6.3.2 ; python_version >= '3.8'
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.6.1 ; python_version >= '3.7'
tzdata==2023.3
tzlocal==5.0.1 ; python_version >= '3.7'
urllib3==2.0.2
validators==0.20.0 ; python_version >= '3.4'
watchdog==3.0.0 ; platform_system != 'Darwin'
wsproto==1.2.0
zipp==3.15.0 ; python_version >= '3.7'
